move test to tests/test_litellm dir
