+++ tests/local_testing/test_router_fallback_handlers.py
@@ -1,5 +1,4 @@
 import asyncio
-import copy
 import os
 import sys
 import time
@@ -317,90 +316,3 @@ async def test_multiple_fallbacks(original_function):
         result._hidden_params["api_base"]
         == "https://exampleopenaiendpoint-production.up.railway.app/"
     )
-
-
-@pytest.mark.asyncio
-async def test_fallback_request_object_not_mutated_for_vision_models():
-    """
-    Tests that when a call to a vision model fails and falls back to another,
-    the `messages` object in the request is not mutated.
-    """
-    model_list = [
-        {
-            "model_name": "openai-vision-fail",
-            "litellm_params": {
-                "model": "openai/gpt-4-vision-preview",
-                "api_key": "bad-key",
-            },
-        },
-        {
-            "model_name": "gemini-vision-success",
-            "litellm_params": {
-                "model": "gemini/gemini-pro-vision",
-                "api_key": os.getenv("GEMINI_API_KEY"),
-            },
-        },
-    ]
-
-    router = Router(
-        model_list=model_list,
-        fallbacks=[{"openai-vision-fail": ["gemini-vision-success"]}],
-        num_retries=0,
-    )
-
-    original_messages = [
-        {
-            "role": "user",
-            "content": [
-                {"type": "text", "text": "What's in this image?"},
-                {
-                    "type": "image_url",
-                    "image_url": {
-                        "url": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mP8z8BQDwAEhQGAhKmMIQAAAABJRU5ErkJggg=="
-                    },
-                },
-            ],
-        }
-    ]
-
-    messages_for_call = copy.deepcopy(original_messages)
-
-    with patch('litellm.router.litellm.acompletion', new_callable=AsyncMock) as mock_acompletion:
-
-        captured_messages_on_fallback = None
-
-        async def side_effect(*args, **kwargs):
-            nonlocal captured_messages_on_fallback
-            model = kwargs.get('model')
-            messages_arg = kwargs.get('messages') 
-
-            if "openai" in model:
-                if messages_arg and isinstance(messages_arg, list):
-                    for message in messages_arg:
-                        if message.get('role') == 'user' and isinstance(message.get('content'), list):
-                            for content_item in message['content']:
-                                if content_item.get('type') == 'image_url':
-                                    # Simulate adding a provider-specific key or modifying structure
-                                    content_item['image_url']['openai_processed_field'] = True
-                                    break 
-                
-                raise litellm.exceptions.AuthenticationError(message="Invalid API key", llm_provider="openai", model=model)
-            elif "gemini" in model:
-                captured_messages_on_fallback = copy.deepcopy(messages_arg)
-                return litellm.ModelResponse(
-                    id="chatcmpl-123",
-                    choices=[litellm.Choices(finish_reason="stop", index=0, message=litellm.Message(content="hello", role="assistant"))],
-                    model="gemini/gemini-pro-vision",
-                )
-
-        mock_acompletion.side_effect = side_effect
-
-        response = await router.acompletion(
-            model="openai-vision-fail",
-            messages=messages_for_call,
-        )
-
-        assert response is not None
-        assert captured_messages_on_fallback is not None
-        assert captured_messages_on_fallback == original_messages
-        assert messages_for_call == original_messages
